Moving factories into _client_factory file
